iron/common and its tests: clean under ruff and pyright
ruff's fixes (import order, docstring closings, four unused imports), with
the one it got wrong undone: iron/common/__init__ imports image.artifacts
before design on purpose, or design's import of image completes a cycle
back into design before DesignGenerator exists.

pyright's findings, by kind: "None until prepared" state narrowed where it
is read (a compiled design's cache entry through one accessor,
jit_compile.cache_entry; a placed arena; a linked mode; a bound device);
Transfers declares the surface a concrete sequence supplies, so the
derived sequence type-checks against it; ElementwiseOperator is generic
over ElementwiseOverlay, not Overlay, so its template reads its own
fields; pyxrt and the XRT tensor types import under TYPE_CHECKING for the
checker and in a try for the host stacks without them; _plus takes an i32
IntegerType rather than the union np_dtype_to_mlir_type returns; the
verify tolerance is judged through a fresh local, since assigning to the
parameter loses its narrowing when Tolerance is Unknown. Upstream typing
gaps are worked around inline, each named: Device.resolve() annotated
-> None, region_op annotating its decorator as the op, _aie.pyi without
get_target_model. The tests assert the Optionals they unpack. E741: the
Operator's overlay TypeVar is OV, not O.

+++ iron/operators/flm/gemm/reference.py
@@ -60,7 +60,7 @@ def reference(input_a, input_b, epilogue=Epilogue.NONE, clamp=None):
     ``torch.sigmoid`` can reproduce. Tolerances have to absorb that part.
     """
     out_dtype = input_a.dtype
-    C = np.matmul(
-        input_a.astype(np.float32), input_b.astype(np.float32)
-    ).astype(out_dtype)
+    C = np.matmul(input_a.astype(np.float32), input_b.astype(np.float32)).astype(
+        out_dtype
+    )
     return apply_epilogue(C, epilogue, clamp)
